[Input] Extract helpers to their own module
This allows importing FormControl without pulling in the complete Input and Textarea modules

Fixes #12656
