+++ packages/material-ui/src/Input/Input.test.js
@@ -5,7 +5,7 @@ import { spy } from 'sinon';
 import { createShallow, createMount, getClasses, unwrap } from '../test-utils';
 import InputAdornment from '../InputAdornment';
 import Textarea from './Textarea';
-import Input, { hasValue, isFilled } from './Input';
+import Input from './Input';
 
 describe('<Input />', () => {
   let shallow;
@@ -430,41 +430,6 @@ describe('<Input />', () => {
       const handleRef = spy();
       mount(<Input multiline inputRef={handleRef} />);
       assert.strictEqual(handleRef.callCount, 1);
-    });
-  });
-
-  describe('hasValue', () => {
-    ['', 0].forEach(value => {
-      it(`is true for ${value}`, () => {
-        assert.strictEqual(hasValue(value), true);
-      });
-    });
-
-    [null, undefined].forEach(value => {
-      it(`is false for ${value}`, () => {
-        assert.strictEqual(hasValue(value), false);
-      });
-    });
-  });
-
-  describe('isFilled', () => {
-    [' ', 0].forEach(value => {
-      it(`is true for value ${value}`, () => {
-        assert.strictEqual(isFilled({ value }), true);
-      });
-
-      it(`is true for SSR defaultValue ${value}`, () => {
-        assert.strictEqual(isFilled({ defaultValue: value }, true), true);
-      });
-    });
-    [null, undefined, ''].forEach(value => {
-      it(`is false for value ${value}`, () => {
-        assert.strictEqual(isFilled({ value }), false);
-      });
-
-      it(`is false for SSR defaultValue ${value}`, () => {
-        assert.strictEqual(isFilled({ defaultValue: value }, true), false);
-      });
     });
   });
